Disabled powersafe
This is a workaround for a bug where the adapter may keep sending Null
frames to indicate that it has entered the standby state, when the
kernel module is removed forcefully (e.g. by killing the VM when the WIFI
device is passed through as an USB device).

A clean solution for this problem would be to implement a watchdog timer
in the firmware and stop sending the control frames when it expires,
if the connection to the kernel module is broken.